+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Fixed: Firing sequence of `common_toolkit_loaded` hook
 - Added: `ctk_environment` filter
 - Added: Ability to change or remove Howdy from admin bar
-- Added: Ability to change login errors
+- Added: Ability to change/remove login errors
 - Added: Ability to cache JSON config file
 - Added: Ability to disable WordPress core, plugin and/or theme updates
 - Added: Ability to modify or disable WordPress heartbeat